backend/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
attrs==25.4.0
azure-core==1.36.0
azure-identity==1.15.0
bcrypt==4.1.3
black==25.9.0
boto3==1.40.59
botocore==1.40.59
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
cmdstanpy==1.3.0
contourpy==1.3.3
coverage==7.11.0
cryptography==46.0.3
cx_Oracle==8.3.0
cycler==0.12.1
Cython==0.29.37
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
# emergentintegrations==0.1.0
et_xmlfile==2.0.0
fastapi==0.110.1
fastuuid==0.14.0
filelock==3.20.0
flake8==7.3.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-genai==1.47.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.15.1
hdbscan==0.8.33
hf-xet==1.2.0
holidays==0.83
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.0.1
idna==3.11
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jq==1.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.12.0
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
litellm==1.79.0
llvmlite==0.45.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mccabe==0.7.0
mcp==1.20.0
mdurl==0.1.2
ml_dtypes==0.5.3
motor==3.3.1
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.10.1
networkx==3.5
numba==0.62.1
numpy==1.26.4
oauthlib==3.3.1
openai==1.99.9
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
patsy==1.0.2
pillow==12.0.0
platformdirs==4.5.0
plotly==6.3.1
pluggy==1.6.0
propcache==0.4.1
prophet==1.2.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
PyMySQL==1.1.1
pynndescent==0.5.13
pyodbc==5.2.0
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.14.0
s5cmd==0.2.0
scikit-image==0.25.2
scikit-learn==1.3.2
scipy==1.11.4
shap==0.49.1
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
sse-starlette==3.0.3
stanio==0.5.1
starlette==0.37.2
statsmodels==0.14.0
stripe==13.1.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.5
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.0
xgboost==3.1.1
xlrd==2.0.2
yarl==1.22.0
zipp==3.23.0
scikit-learn==1.3.2
scipy==1.11.4
lightgbm==4.6.0